test(codex): cover attempt-owned tool cleanup
